Move histo and plotter instances to rate configuration
