+++ cfg/rate.yaml
@@ -72,14 +72,15 @@ collections:
         - nugun_alleta_pu200
       plotters:
         # - !!python/name:cfg.rate.eg_emu_plotters
+        - !!python/name:cfg.rate.eg_menuSta_plotters
         - !!python/name:cfg.rate.eg_menuCTl2_plotters
-        - !!python/name:cfg.rate.eg_menuCTl2_ell_plotters
+        # - !!python/name:cfg.rate.eg_menuCTl2_ell_plotters
       htc_jobflavor:
         microcentury
-      priorities:
-        nugun_alleta_pu200: 6
-      events_per_job:
-        nugun_alleta_pu200: 20000
+      # priorities:
+      #   nugun_alleta_pu200: 6
+      # events_per_job:
+      #   nugun_alleta_pu200: 20000
 
   eg_rate_counter_menu:
       file_label:
