+++ python/l1THistos.py
@@ -452,103 +452,6 @@ class DigiHistos(BaseHistos):
         # rnp.fill_hist(self.h_simenergy, digis.simenergy)
 
 
-class RateHistos(BaseHistos):
-    def __init__(self, name, root_file=None, debug=False):
-        if not root_file:
-            self.h_norm = bh.TH1F(f'{name}_norm', '# of events', 1, 1, 2)
-            self.h_pt = bh.TH1F(f'{name}_pt', 'rate above p_{T} thresh.; p_{T} [GeV]; rate [kHz];', 100, 0, 100)
-            # self.h_ptVabseta = bh.TH2F(name+'_ptVabseta', 'Candidate p_{T} vs |#eta|; |#eta|; p_{T} [GeV];', 34, 1.4, 3.1, 100, 0, 100)
-
-        BaseHistos.__init__(self, name, root_file, debug)
-
-        if root_file is not None:
-            for attr_1d in [attr for attr in dir(self) if (attr.startswith('h_') and 'TH1' in getattr(self, attr).ClassName())]:
-                setattr(self, f'{attr_1d}_graph', GraphBuilder(self, attr_1d))
-
-        if root_file is not None:
-            self.normalize(2760.0*11246/1000)
-            # self.h_simenergy = bh.TH1F(name+'_energy', 'Digi sim-energy (GeV)', 100, 0, 2)
-
-    def fill(self, data):
-        # print(self.h_pt.axes[0])
-        pt_max = ak.max(data.pt, axis=1)
-        for thr,bin_center in zip(self.h_pt.axes[0].edges, self.h_pt.axes[0].centers, strict=False):
-        # for thr,bin_center in zip(self.h_pt.axes[0].edges[1:], self.h_pt.axes[0].centers):
-            self.h_pt.fill(bin_center, weight=ak.sum(pt_max>=thr))
-
-        # for ptf in range(0, int(pt)+1):
-        #     self.h_pt.Fill(ptf)
-        # self.h_ptVabseta.Fill(abs(eta), pt)
-
-    def fill_norm(self, many=1):
-        # print (f' fill rate norm: {many}')
-        self.h_norm.fill(1, weight=many)
-
-    def normalize(self, norm):
-        nev = self.h_norm.GetBinContent(1)
-        if(nev != norm):
-            print(f'normalize # ev {nev} to {norm}')
-            self.h_norm.Scale(norm/nev)
-            self.h_pt.Scale(norm/nev)
-            # self.h_ptVabseta.Scale(norm/nev)
-
-
-class RateHistoCounter(BaseHistos):
-    def __init__(self, name, root_file=None, debug=False):
-        if not root_file:
-            self.h_norm = bh.TH1F(f'{name}_norm', '# of events', 1, 1, 2)
-            self.h_rate = bh.TH1F(f'{name}_rate', '# passing events; rate [kHz];',  1, 1, 2)
-
-        BaseHistos.__init__(self, name, root_file, debug)
-
-        if root_file is not None or True:
-            for attr_1d in [attr for attr in dir(self) if (attr.startswith('h_') and 'TH1' in getattr(self, attr).ClassName())]:
-                setattr(self, f'{attr_1d}_graph', GraphBuilder(self, attr_1d))
-
-        if root_file is not None:
-            self.normalize(31000)
-            # self.h_simenergy = bh.TH1F(name+'_energy', 'Digi sim-energy (GeV)', 100, 0, 2)
-
-    def fill(self, count):
-        self.h_rate.Fill(1, count)
-
-    def fill_norm(self, many=1):
-        # print (f' fill rate norm: {many}')
-        self.h_norm.Fill(1, many)
-
-    def normalize(self, norm):
-        nev = self.h_norm.GetBinContent(1)
-        # print(f' .      # ev: {nev}')
-        if(nev != norm):
-            print(f'normalize to {norm}')
-            self.h_norm.Scale(norm/nev)
-            self.h_rate.Scale(norm/nev)
-
-
-class SingleObjRateHistoCounter(RateHistoCounter):
-    def __init__(self, name, root_file=None, debug=False):
-        RateHistoCounter.__init__(self, name, root_file, debug)
-
-    def fill(self, df):
-        # print(f' .  # of surviving entries: ')
-        # print(df['pt'].groupby(level='entry', group_keys=False).nlargest(n=1).count())
-        RateHistoCounter.fill(self, df['pt'].groupby(level='entry', group_keys=False).nlargest(n=1).count())
-
-
-class DoubleObjRateHistoCounter(RateHistoCounter):
-    def __init__(self, name, root_file=None, debug=False):
-        RateHistoCounter.__init__(self, name, root_file, debug)
-
-    def has_unique_pairs(entry):
-        unique_pairs = [(l1, l2) for l1 in entry.loc[entry.index.get_level_values('leg') == 0].index for l2 in entry.loc[entry.index.get_level_values('leg') == 1].index if ((l1[2] != l2[2]) and (l1[1] != l2[1]))]
-        return len(unique_pairs) > 0
-
-    def fill(self, df):
-        # print(f' .  # of surviving entries: ')
-        # print(df.groupby(level='entry').filter(DoubleObjRateHistoCounter.has_unique_pairs).index.unique('entry').size)
-        RateHistoCounter.fill(self, df.groupby(level='entry').filter(DoubleObjRateHistoCounter.has_unique_pairs).index.unique('entry').size)
-
-
 class TCHistos(BaseHistos):
     def __init__(self, name, root_file=None, debug=False):
         if not root_file:
